MybatisでINSERT、ロガー定義簡易化

+++ kotlin-backend/src/main/kotlin/com/sanyu_i/sisdiv3/n_study/ntsa/kotlinbackend/api/sample/dataset/Sample.kt
@@ -1,3 +1,3 @@
 package com.sanyu_i.sisdiv3.n_study.ntsa.kotlinbackend.api.sample.dataset;
 
-data class Sample(var id: Int, var name: String)
+data class Sample(var id: Long, var name: String)
